mobile navbar
seperate the mobile version of the navbar into its own component,
transfer all logic and styles into a seperate file for the mobile
navbar.
