+++ src/components/NavBar/NavBar.module.css
@@ -14,80 +14,6 @@
   align-items: center;
 }
 
-.mobileNavMenu {
-  --borderColor: rgba(228, 228, 228, 0.3);
-  background-color: var(--background);
-  position: absolute;
-  top: 4.3em;
-  left: 0;
-  width: 100%;
-  border-top: 1px solid var(--borderColor);
-  box-shadow: 0 6px 50px -5px rgba(0, 0, 0, 0.5);
-  transform-origin: top center;
-  transform: rotateX(-90deg);
-  opacity: 0.3;
-  visibility: hidden;
-  transition: var(--transition);
-}
-
-.mobileNavMenu.showMenu {
-  opacity: 1;
-  transform: rotateX(0);
-  visibility: visible;
-}
-
-.mobileNavList {
-  color: var(--light-text);
-}
-
-.mobileNavItem {
-  font-size: 1.4rem;
-  font-family: inherit;
-  text-align: right;
-  padding: 1em;
-  padding-right: 0;
-  border-bottom: 1px solid var(--borderColor);
-  position: relative;
-}
-
-.mobileNavItemContainer {
-  composes: container from '../../styles/utils.module.css';
-}
-
-.mobileNavItem a {
-  color: var(--light-text);
-  width: 100%;
-}
-
-.mobileNavItem a::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
-
-.mobileNavItem:last-of-type a {
-  color: var(--primary);
-}
-
-.mobileNavItem a:focus,
-.mobileNavItem a:active {
-  color: var(--primary);
-}
-
-.showMenuFixed {
-  display: fixed;
-  top: 0;
-  left: 0;
-}
-
-.mobileNavButton,
-.mobileNavButtonClose {
-  background: transparent;
-}
-
 .navDesktop {
   color: var(--light-text);
   display: none;
@@ -120,20 +46,7 @@
 }
 
 @media screen and (min-width: 768px) {
-  .mobileNavMenu {
-    display: none;
-  }
-
-  .mobileNav {
-    display: none;
-  }
-
   .navDesktop {
     display: block;
-  }
-
-  .mobileNavButton,
-  .mobileNavButtonClose {
-    display: none;
   }
 }
